DEX-23493 | cleaned up code in ghosti
