+++ ghosti/constants.js
@@ -2,6 +2,20 @@ const SNAPSHOTS_SUITE_ID = '64c8d884960593b38bb68331';
 const PATH_TO_BLOCKS = 'sidekick/blocks';
 const LOCAL_BLOCKS_PATH = '_src-lp/blocks';
 const FETCH_TIMEOUT = 1000 * 60 * 5; // 5 minutes
+// todo add sidekick config for those
+const EXCLUDED_SNAPSHOT_BLOCKS = [
+  'aem-banner',
+  'aem-two-cards',
+  'b-industry-recognition',
+  'cards',
+  'columns-two',
+  'fragment',
+  'header',
+  'lp-custom',
+  'support',
+  'tos',
+  'video',
+];
 
 function logError(message) {
   console.log('\x1b[31m%s\x1b[0m', message);
@@ -15,6 +29,7 @@ module.exports = {
   PATH_TO_BLOCKS,
   LOCAL_BLOCKS_PATH,
   FETCH_TIMEOUT,
+  EXCLUDED_SNAPSHOT_BLOCKS,
   logError,
   logSuccess
 }
